refactor: update references to EN 1993-1-1:2022 in formulas 8.31, 8.32, and 8.33
